# openapi.yaml
openapi: 3.1.0
...
info:
  version: 0.1.0
  title: Quack API
  description: Local-first DuckDB-backed API for file ingestion and SQL execution.
servers:
  - url: http://localhost:3001
paths:
...
  /documents:
    post:
      summary: Create document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDocumentRequest'
      responses:
        '200':
          description: Document created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: List documents
      responses:
        '200':
          description: Array of documents
          content:
            application/json:
              schema:
                type: object
                properties:
                  documents:
                    type: array
                    items:
                      $ref: '#/components/schemas/Document'
  /documents/{id}:
    get:
      summary: Get document by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Document details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentWithCells'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Update document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDocumentRequest'
      responses:
        '200':
          description: Document updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentWithCells'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Document deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/{id}/export:
    post:
      summary: Export document to portable .quackdb file
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExportRequest'
      responses:
        '200':
          description: Export successful
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid input or dataMode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/import:
    post:
      summary: Import document from .quackdb file
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
              required:
                - file
      responses:
        '200':
          description: Import successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentWithCells'
        '400':
          description: Invalid file or format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    Document:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        markdown:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - created_at
        - updated_at
    DocumentCell:
      type: object
      properties:
        id:
          type: string
        document_id:
          type: string
        cell_index:
          type: integer
        cell_type:
          type: string
          enum: [sql, markdown]
        sql_text:
          type: string
          nullable: true
        markdown_text:
          type: string
          nullable: true
        chart_config:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
      required:
        - id
        - document_id
        - cell_index
        - cell_type
        - created_at
    DocumentWithCells:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        markdown:
          type: string
          nullable: true
        cells:
          type: array
          items:
            $ref: '#/components/schemas/DocumentCell'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - cells
        - created_at
        - updated_at
    CreateDocumentRequest:
      type: object
      properties:
        name:
          type: string
        markdown:
          type: string
      required:
        - name
    UpdateDocumentRequest:
      type: object
      properties:
        name:
          type: string
        markdown:
          type: string
    ExportRequest:
      type: object
      properties:
        dataMode:
          type: string
          enum: [none, query-results, referenced-tables, full-db]
          default: query-results
      required:
        - dataMode
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error